fixes: Query members, service wasn't advertised
Two names in the Query service message were renamed to better fit their
actual meaning.

A bug where the service wasn't actually advertised was fixed. It was due
to my misunderstanding of the semantics of the ros::ServiceServer
object. Basically, the service gets torn down as soon as the object is
destroyed; if it isn't kept around in a variable, such that the object
is still alive at the ros::spin() call, it won't ever be visible to the
rest of the ROS system.